metrics_exporter/requirements.txt
mysql-connector-python==8.1.0
flask==2.3.3
prometheus-client==0.17.1
